requirements.txt
flask
flask-bootstrap
flask-wtf
python-dotenv
black
